feat(org): Execution Groups
Add entity `org_execgroup`, which is a part of organizational structure, is `unitType` is `EXECGROUP`.
The detail entity `org_execgroupmember` contains its members, which are `org_staffunit`.
Business value of this entity is ability to use a group of staff units at once, for example for task assignment,
like parallel approvals.
Added a navigational menu item for this dictionary in `org_desktop`.