fix(ub-pub): AsyncConnection to parse errors for UB in prod mode
UB in production mode return just error.  UB in dev mode return JSON object.  That was not taken into account.

Fix is to require either double-quote sign OR begin or line at match start.  And similarly at match end - require quote OR end of line.

So, changed in original:
const PARSE_ERROR_MESSAGE_RE = /"<<<(.*?)>>>(?:\|(\[[^\]]*\]))?"/

to

const PARSE_ERROR_MESSAGE_RE = /(?:^|")<<<(.*?)>>>(?:\|(\[[^\]]*\]))?(?:$|")/

The `(?:^|")` is non-matching group with alternation - the ^ (start of line) and symbol ".

Also, in `JSON.parse(msgUnparsed ....)` part (like 63), we have to analyze if string started with quote or not.