Introduce the `uiSettingsStorage` module for setting/getting/clearing UI settings
It is recommended to use this module instead of the `localStorage` for persisting
some settings on UI. These settings are cleared on the `Reset GUI Settings` action
and emit the `portal:resetGUI` event. Also, it is added to `Vue.prototype.$uiSettingsStorage`.

Migrate from using the `localStorage` to `uiSettingsStorage` in appropriate places